[IMP] purchase_order_line_sequence: remove dependency from stock_picking_line_sequence
