+++ purchase_order_line_sequence/__manifest__.py
@@ -5,8 +5,7 @@
 
 {
     "name": "Purchase Order Line Sequence",
-    "summary": "Adds sequence to PO lines and propagates it to"
-    "Invoice lines and Stock Moves",
+    "summary": "Adds sequence to PO lines and propagates it to Invoice lines",
     "version": "18.0.1.0.1",
     "category": "Purchase Management",
     "author": "Camptocamp, "
@@ -15,8 +14,7 @@
     "Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/purchase-workflow",
     "depends": [
-        "purchase_stock",
-        "stock_picking_line_sequence",
+        "purchase",
     ],
     "data": [
         "views/purchase_view.xml",
