+++ purchase_order_line_sequence/readme/DESCRIPTION.md
@@ -1,2 +1,7 @@
-The sequence in PO line is propagated to the Stock moves. The sequence
-number appears in the PO form view and in the report.
+Adds a visible line-number sequence to purchase orders. The sequence
+number appears in the PO form view, in the purchase order/quotation
+reports, and is propagated to the related vendor bill lines.
+
+When `stock_picking_line_sequence` is also installed, the glue module
+`purchase_stock_line_sequence` is auto-installed and propagates this
+sequence to the related stock moves as well.
